CON-1077 - Move GraalVM patch to new public fork
Moved the graal module from the conclave-sdk repo to this new repo.
Fixed a bug in the script cloneAndPatch.sh to ensure that the same git commit is always used after cloning graal.
Other changes were applied to fix the issues that were caused after moving the graal module to the new repo.